split nc into more files
